Switch from SDL2 to SDL3

+++ int.inc.h
@@ -1,5 +1,5 @@
 //extern volatile uint32_t maincount;
-#define maincount (SDL_GetTicks())
+#define maincount ((uint32_t)SDL_GetTicks())
 
 //extrn i8_init:near, i8_done:near
 //extrn newhandler:near, add_time:near, reset_timestep:near
